ksm002 update tab index

+++ nts.uk/uk.at/at.web/nts.uk.at.web/src/main/webapp/view/ksm/002/a/ksm002.a.vm.ts
@@ -18,7 +18,7 @@ module ksm002.a.viewmodel {
         yearMonthPicked: KnockoutObservable<number>;
         cssRangerYM: any;
         optionDates: KnockoutObservableArray<OptionalDate>;
-        firstDay: number;
+        firstDay: KnockoutObservable<number>;
         yearMonth: KnockoutObservable<number>;
         startDate: number;
         endDate: number;
@@ -41,13 +41,13 @@ module ksm002.a.viewmodel {
             self.cssRangerYM = {};
             self.optionDates = ko.observableArray([]);
 
-            self.firstDay = 3;
+            self.firstDay = ko.observable(0);
             self.startDate = 1;
             self.endDate = 31;
             self.workplaceId = ko.observable("0");
             self.workplaceName = ko.observable("");
             self.eventDisplay = ko.observable(true);
-            self.eventUpdatable = ko.observable(true);
+            self.eventUpdatable = ko.observable(false);
             self.holidayDisplay = ko.observable(true);
             self.cellButtonDisplay = ko.observable(true);
 
@@ -56,8 +56,10 @@ module ksm002.a.viewmodel {
                     self.openKsm002EDialog(date);
                 },
                 cellClick: function(date) {
+                    console.time("ksm002_CellClick");
                     let param: IData = { date: date, selectable: _.map(self.boxItemList(), 'id'), selecteds: self.selectedIds() };
                     self.setSpecificItemToSelectedDate(param);
+                    console.timeEnd("ksm002_CellClick");
                 }
             });
             //Side bar tab change
@@ -74,6 +76,7 @@ module ksm002.a.viewmodel {
             //Change Month 
             self.yearMonthPicked.subscribe(function(value) {
                 let arrOptionaDates: Array<OptionalDate> = [];
+                console.time("ksm002_ChangeMonth");
                 self.getDataToOneMonth(value).done(function(arrOptionaDates) {
                     self.optionDates(arrOptionaDates);
                     self.optionDates.valueHasMutated();
@@ -82,6 +85,7 @@ module ksm002.a.viewmodel {
                     } else {
                         self.isNew(true);
                     }
+                    console.timeEnd("ksm002_ChangeMonth");
                 })
             })
         }
@@ -96,19 +100,17 @@ module ksm002.a.viewmodel {
             let arrOptionaDates: Array<OptionalDate> = [];
             service.getSpecificDateByIsUse(isUse).done(function(lstSpecifiDate: any) {
                 if (lstSpecifiDate.length > 0) {
+                    console.time("ksm002_LoadData")
                     //getAll SpecDate
                     self.getAllSpecDate();
                     //Set Start Day of Company
-                    //self.firstDay = 3;
-                    //Test 
-                    //nts.uk.characteristics.save('IndividualStartDay', 0);
                     nts.uk.characteristics.restore('IndividualStartDay').done(function (data) { 
                         if(nts.uk.util.isNullOrEmpty(data)){
                             self.getComStartDay().done(function(startDay: number) {
-                                self.firstDay = startDay;
+                                self.firstDay(startDay);
                             });
                         } else {
-                            self.firstDay = data;
+                            self.firstDay(data);
                         }
                     });
                     //set parameter to calendar
@@ -116,13 +118,13 @@ module ksm002.a.viewmodel {
                     _.forEach(lstSpecifiDate, function(item) {
                         lstBoxCheck.push(new BoxModel(item.specificDateItemNo, item.specificName));
                     });
-                    _.orderBy(lstBoxCheck, ['id'], ['desc']);
-                    self.boxItemList(lstBoxCheck);
+                    self.boxItemList(_.orderBy(lstBoxCheck, ['id'], ['asc']));
                     //Set data to calendar
                     self.getDataToOneMonth(self.yearMonthPicked()).done(function(arrOptionaDates) {
                         if (arrOptionaDates.length > 0) {
                             self.optionDates(arrOptionaDates);
                             self.isNew(false);
+                            console.timeEnd("ksm002_LoadData");
                         }
                     })
                 } else {
@@ -215,14 +217,12 @@ module ksm002.a.viewmodel {
             
             if (nts.uk.util.isNullOrEmpty(selectedOptionalDate)) {
                 self.optionDates.push(new OptionalDate(selectedDate, self.getNamefromSpecId(param.selecteds), param.selecteds));
-                //self.optionDate.valueHasMutated();
             } else {
                 self.optionDates.remove(selectedOptionalDate);
                 selectedOptionalDate.start = selectedDate;
                 selectedOptionalDate.listId = param.selecteds;
                 selectedOptionalDate.listText = self.getNamefromSpecId(param.selecteds);
                 self.optionDates.push(selectedOptionalDate);
-                //self.optionDate.valueHasMutated();
             }
         }
         /**
@@ -230,9 +230,12 @@ module ksm002.a.viewmodel {
          */
         Insert(lstComSpecificDateCommand: Array<ICompanySpecificDateCommand>) {
             var self = this;
+            nts.uk.ui.block.invisible();
             service.insertComSpecificDate(lstComSpecificDateCommand).done(function(res: Array<any>) {
                 nts.uk.ui.dialog.info({ messageId: "Msg_15" }).then(function() {
                     self.start();
+                    console.timeEnd("ksm002_Register");
+                    nts.uk.ui.block.clear();
                 });
             }).fail(function(res) {
                 nts.uk.ui.dialog.alertError(res.message).then(function() { nts.uk.ui.block.clear(); });
@@ -245,19 +248,27 @@ module ksm002.a.viewmodel {
         DeleteOneMonth() {
             let dfd = $.Deferred<any>();
             var self = this;
-            //delete
-            service.deleteComSpecificDate({ yearMonth: self.yearMonthPicked().toString() }).done(function(res: any) {
-                nts.uk.ui.dialog.info({ messageId: "Msg_16" }).then(function() {
-                    //Set dataSource to Null
-                    self.optionDates([]);
-                    self.isNew(true);
+            nts.uk.ui.dialog.confirm({ messageId: 'Msg_18' }).ifYes(function() {
+                nts.uk.ui.block.invisible();
+                //delete
+                console.time("ksm002_DeleteOneMonth");
+                service.deleteComSpecificDate({ yearMonth: self.yearMonthPicked().toString() }).done(function(res: any) {
+                    nts.uk.ui.dialog.info({ messageId: "Msg_16" }).then(function() {
+                        //Set dataSource to Null
+                        self.optionDates([]);
+                        self.isNew(true);
+                        console.timeEnd("ksm002_DeleteOneMonth");
+                    });
+                    nts.uk.ui.block.clear();
+                    dfd.resolve();
+                }).fail(function(res) {
+                    nts.uk.ui.dialog.alertError(res.message).then(function() { nts.uk.ui.block.clear(); });
+                    dfd.reject();
                 });
-                dfd.resolve();
-            }).fail(function(res) {
-                nts.uk.ui.dialog.alertError(res.message).then(function() { nts.uk.ui.block.clear(); });
-                dfd.reject();
+                return dfd.promise();
+            }).ifNo(function() {
+                // do nothing           
             });
-            return dfd.promise();
         }
 
         /**
@@ -271,24 +282,30 @@ module ksm002.a.viewmodel {
                 nts.uk.ui.dialog.alertError({ messageId: "Msg_139" });
             } else {
                 //New case
+                console.time("ksm002_Register");
                 if (self.isNew()) {
                     self.Insert(self.convertToCommand())
                 } else {
                     //Update case
                     if (_.flattenDepth(_.map(self.optionDates(), 'listId')).length > 0) {
                         //update has data
+                        nts.uk.ui.block.invisible();
                         service.updateComSpecificDate(self.convertToCommand()).done(function(res: Array<any>) {
                             nts.uk.ui.dialog.info({ messageId: "Msg_15" }).then(function() {
                                 self.start();
+                                nts.uk.ui.block.clear();
+                                console.timeEnd("ksm002_Register");
                             });
                         }).fail(function(res) {
                             nts.uk.ui.dialog.alertError(res.message).then(function() { nts.uk.ui.block.clear(); });
                         });
                     } else {
                         //Delete all OLD data
+                        nts.uk.ui.block.invisible();
                         service.deleteComSpecificDate({ yearMonth: self.yearMonthPicked().toString() }).done(function(res: any) {
                             nts.uk.ui.dialog.info({ messageId: "Msg_15" }).then(function() {
                                 self.isNew(true);
+                                nts.uk.ui.block.clear();
                             });
                         }).fail(function(res) {
                             nts.uk.ui.dialog.alertError(res.message).then(function() { nts.uk.ui.block.clear(); });
@@ -381,7 +398,7 @@ module ksm002.a.viewmodel {
                     endDate: Number(moment(self.yearMonthPicked().toString(), 'YYYYMM').endOf('month').format('YYYYMMDD'))
                 });
             nts.uk.ui.windows.sub.modal("/view/ksm/002/d/index.xhtml", { title: "割増項目の設定", dialogClass: "no-close" }).onClosed(function() {
-
+                self.start();
             });
         }
